src/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
PyYAML==6.0.2
httpx==0.28.1
openai==1.58.1
numpy==2.2.1
azure-identity==1.19.0
azure-search-documents==11.6.0b8
azure-storage-blob==12.24.1
azure-cosmos==4.9.0
pyodbc==5.2.0